Neu: * Option zum Deaktivieren der Tastatureingabe (F2 für Kontaktsuche) * Option zur Deaktivierung der E-Mailüberrüfung für Ribbion Dial Button (#45) Optimierung Kontaktsuche * Aktivierung mittels F2 * Focus auf Eingabe TextBox * Suchtasks werden beendet, wenn schnell getippt wird.

+++ FBoxDial/Formulare/Kontaktsuche/ViewModel/KontaktsucheViewModel.vb
@@ -1,14 +1,16 @@
-﻿Public Class KontaktsucheViewModel
+﻿Imports System.Threading.Tasks
+Imports Microsoft.Office.Interop.Outlook
+Public Class KontaktsucheViewModel
     Inherits NotifyBase
     'Private Property NLogger As Logger = LogManager.GetCurrentClassLogger
     Private Property DatenService As IDataKontaktsuche
-    'Private Property DialogService As IDialogService
 
 #Region "ICommand"
 
 #End Region
 
 #Region "Properties"
+    Private SearchTask As Task(Of List(Of ContactItem))
 
     Private _SearchText As String
     Public Property SearchText As String
@@ -66,14 +68,19 @@
         End If
 
         If SearchText.IsNotStringNothingOrEmpty Then
-            ' Führe die Suche durch
-            Dim NeuesErgebnis As List(Of ContactItemViewModel) = (Await DatenService.KontaktSuche(SearchText)).Select(Function(K) New ContactItemViewModel With {.OlKontakt = K}).ToList
 
-            ' Füge alle neuen Kontakte hinzu
-            Ergebnisse.AddRange(NeuesErgebnis.Except(Ergebnisse))
+            SearchTask = DatenService.KontaktSuche(SearchText)
 
-            ' Entferne alle nicht mehr passenden Kontakte
-            Ergebnisse.RemoveRange(Ergebnisse.Except(NeuesErgebnis))
+            If Not SearchTask.IsCanceled Then
+                ' Führe die Suche durch
+                Dim NeuesErgebnis As List(Of ContactItemViewModel) = (Await SearchTask).Select(Function(K) New ContactItemViewModel With {.OlKontakt = K}).ToList
+                ' Füge alle neuen Kontakte hinzu
+                Ergebnisse.AddRange(NeuesErgebnis.Except(Ergebnisse))
+
+                ' Entferne alle nicht mehr passenden Kontakte
+                Ergebnisse.RemoveRange(Ergebnisse.Except(NeuesErgebnis))
+            End If
+
         Else
             Ergebnisse.Clear()
         End If
